+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 3.3.1.0
 
 - Install packs from Curse through dragging and dropping the Install button
+- Sdd in servers tab to manage servers created through the launcher